+++ src/sagan/Reactor.fs
@@ -24,7 +24,7 @@ module internal Reactor =
     { send = (fun b -> r.send (g b)) ; events = r.events |> AsyncSeq.map f }
 
 
-type ChangeFeedPositionTracker<'a> = private {
+type internal ChangeFeedPositionTracker<'a> = private {
   channel :  Reactor<'a>
 }