+++ RELEASE_NOTES.md
@@ -1,6 +1,10 @@
+### 0.0.3
+* Changed `PartitionPosition` to `RangePosition` (without `PartitionId`)
+* Made `ChangeFeedPositionTracker` internal
+
 ### 0.0.2
 * Assembly name fixup
 * Dependencies minimum version numbers
 
-### 0.0.1 - 29.12.2017
+### 0.0.1
 * initial